+++ packages/las-sdk-core/src/client.ts
@@ -9,11 +9,8 @@ import {
   AssetList,
   AuthorizationHeaders,
   AxiosFn,
-  Batch,
-  BatchList,
   ContentType,
   CreateAppClientOptions,
-  CreateBatchOptions,
   CreateDataBundleOptions,
   CreateDatasetOptions,
   CreateDocumentOptions,
@@ -34,7 +31,6 @@ import {
   LasDocumentWithoutContent,
   ListAppClientsOptions,
   ListAssetsOptions,
-  ListBatchesOptions,
   ListDataBundleOptions,
   ListDatasetsOptions,
   ListDocumentsOptions,
@@ -62,7 +58,6 @@ import {
   TransitionType,
   UpdateAppClientOptions,
   UpdateAssetOptions,
-  UpdateBatchOptions,
   UpdateDataBundleOptions,
   UpdateDatasetOptions,
   UpdateDocumentOptions,
@@ -169,7 +164,6 @@ export class Client {
    * @param content Content to POST (base64 string | Buffer)
    * @param contentType MIME type for the document
    * @param options.consentId Id of the consent that marks the owner of the document
-   * @param options.batchId Id of the associated batch
    * @param options.groundTruth List of GroundTruth items representing the ground truth values for the document
    * @returns Document response from REST API
    */
@@ -204,7 +198,6 @@ export class Client {
   /**
    * List documents available for inference, calls the GET /documents endpoint.
    *
-   * @param options.batchId Ids of the batches that contains the documents of interest
    * @param options.consentId Ids of the consents that marks the owner of the document
    * @param options.maxResults Maximum number of results to be returned
    * @param options.nextToken A unique token for each page, use the returned token to retrieve the next page.
@@ -763,63 +756,6 @@ export class Client {
    */
   async updateDataBundle(modelId: string, dataBundleId: string, options: UpdateDataBundleOptions): Promise<DataBundle> {
     return this.makePatchRequest(`/models/${modelId}/dataBundles/${dataBundleId}`, options);
-  }
-
-  /**
-   * @deprecated Use the new {@link Client.createDataset} method instead.
-   * Creates a batch, calls the POST /batches endpoint.
-   *
-   * @param options.name Name of the batch
-   * @param options.description Description of the batch
-   * @returns Batch response from REST API
-   */
-  async createBatch(options: CreateBatchOptions): Promise<Batch> {
-    return this.makePostRequest<Batch>('/batches', options);
-  }
-
-  /**
-   * Updates an batch, calls the PATCH /batches/{batchId} endpoint.
-   *
-   * @deprecated Use the new {@link Client.updateDataset} method instead.
-   * @param batchId Id of the batch
-   * @param options.description Description of batch
-   * @param options.name Name of batch
-   * @returns Batch response from REST API with content
-   */
-  async updateBatch(batchId: string, options: UpdateBatchOptions): Promise<Batch> {
-    return this.makePatchRequest(`/batches/${batchId}`, options);
-  }
-
-  /**
-   * List batches, calls the GET /batches endpoint.
-   *
-   * @deprecated Use the new {@link Client.listDatasets} method instead.
-   * @param options.maxResults Maximum number of results to be returned
-   * @param options.nextToken A unique token for each page, use the returned token to retrieve the next page.
-   * @returns BatchList response from REST API
-   */
-  async listBatches(options?: ListBatchesOptions): Promise<BatchList> {
-    return this.makeGetRequest<BatchList>('/batches', options);
-  }
-
-  /**
-   * Deletes a batch, calls the DELETE /batches/{batchId} endpoint.
-   *
-   * @deprecated Use the new {@link Client.deleteDataset} method instead.
-   * @param batchId Id of the batch
-   * @param deleteDocuments Set to true to delete documents in batch before deleting batch
-   * @returns Batch response from REST API
-   */
-  async deleteBatch(batchId: string, deleteDocuments = false): Promise<Batch> {
-    if (deleteDocuments) {
-      let response = await this.deleteDocuments({ batchId });
-      while (response.nextToken) {
-        // eslint-disable-next-line no-await-in-loop
-        response = await this.deleteDocuments({ batchId, nextToken: response.nextToken });
-      }
-    }
-
-    return this.makeDeleteRequest<Batch>(`/batches/${batchId}`);
   }
 
   /**